+++ memtable-core/src/impls/col.rs
@@ -1,4 +1,4 @@
-use crate::{iter::*, list::*, utils, Position, Table};
+use crate::{iter::*, list::*, utils, Capacity, Position, Table};
 use core::{
     cmp,
     iter::FromIterator,
@@ -68,6 +68,14 @@ impl<T: Default, const COL: usize> Table for FixedColumnTable<T, COL> {
     type Data = T;
     type Row = FixedList<Self::Data, COL>;
     type Column = DynamicList<Self::Data>;
+
+    fn max_row_capacity(&self) -> Capacity {
+        Capacity::Unlimited
+    }
+
+    fn max_column_capacity(&self) -> Capacity {
+        Capacity::Limited(COL)
+    }
 
     fn row_cnt(&self) -> usize {
         self.row_cnt
@@ -140,7 +148,7 @@ impl<T: Default, const COL: usize> Table for FixedColumnTable<T, COL> {
     ///
     /// If you want to remove the cells that are no longer within capacity,
     /// call [`Self::truncate`], which will reset them to their default value.
-    fn set_row_capacity(&mut self, capacity: usize) {
+    fn set_preferred_row_cnt(&mut self, capacity: usize) {
         self.row_cnt = capacity;
     }
 
@@ -153,7 +161,7 @@ impl<T: Default, const COL: usize> Table for FixedColumnTable<T, COL> {
     ///
     /// If you want to remove the cells that are no longer within capacity,
     /// call [`Self::truncate`], which will reset them to their default value.
-    fn set_column_capacity(&mut self, capacity: usize) {
+    fn set_preferred_col_cnt(&mut self, capacity: usize) {
         self.col_cnt = cmp::min(capacity, COL);
     }
 }
@@ -291,12 +299,12 @@ mod tests {
     fn row_cnt_should_be_adjustable() {
         let mut table: FixedColumnTable<usize, 0> = FixedColumnTable::new();
         assert_eq!(table.row_cnt(), 0);
-        table.set_row_capacity(999);
+        table.set_preferred_row_cnt(999);
         assert_eq!(table.row_cnt(), 999);
 
         let mut table: FixedColumnTable<usize, 4> = FixedColumnTable::new();
         assert_eq!(table.row_cnt(), 0);
-        table.set_row_capacity(999);
+        table.set_preferred_row_cnt(999);
         assert_eq!(table.row_cnt(), 999);
     }
 
@@ -304,12 +312,12 @@ mod tests {
     fn col_cnt_should_be_adjustable_up_to_const_max() {
         let mut table: FixedColumnTable<usize, 0> = FixedColumnTable::new();
         assert_eq!(table.col_cnt(), 0);
-        table.set_column_capacity(1);
+        table.set_preferred_col_cnt(1);
         assert_eq!(table.col_cnt(), 0);
 
         let mut table: FixedColumnTable<usize, 4> = FixedColumnTable::new();
         assert_eq!(table.col_cnt(), 0);
-        table.set_column_capacity(5);
+        table.set_preferred_col_cnt(5);
         assert_eq!(table.col_cnt(), 4);
     }
 
@@ -332,16 +340,16 @@ mod tests {
         assert_eq!(table.col_cnt(), 2);
 
         // If we change the capacity to be smaller, cell should respect that
-        table.set_row_capacity(1);
-        table.set_column_capacity(1);
+        table.set_preferred_row_cnt(1);
+        table.set_preferred_col_cnt(1);
         assert_eq!(table.cell(0, 0).as_deref(), Some(&"a"));
         assert_eq!(table.cell(0, 1).as_deref(), None);
         assert_eq!(table.cell(1, 0).as_deref(), None);
         assert_eq!(table.cell(1, 1).as_deref(), None);
 
         // Capacity changes don't actually overwrite anything
-        table.set_row_capacity(2);
-        table.set_column_capacity(2);
+        table.set_preferred_row_cnt(2);
+        table.set_preferred_col_cnt(2);
         assert_eq!(table.cell(0, 0).as_deref(), Some(&"a"));
         assert_eq!(table.cell(0, 1).as_deref(), Some(&"b"));
         assert_eq!(table.cell(1, 0).as_deref(), Some(&"c"));
@@ -370,8 +378,8 @@ mod tests {
         assert_eq!(table.col_cnt(), 2);
 
         // If we change the capacity to be smaller, mut_cell should respect that
-        table.set_row_capacity(1);
-        table.set_column_capacity(1);
+        table.set_preferred_row_cnt(1);
+        table.set_preferred_col_cnt(1);
         assert!(table.mut_cell(0, 0).is_some());
         assert!(table.mut_cell(0, 1).is_none());
         assert!(table.mut_cell(1, 0).is_none());
@@ -432,8 +440,8 @@ mod tests {
     #[test]
     fn remove_cell_should_respect_virtual_boundaries() {
         let mut table = FixedColumnTable::from(vec![[1, 2], [3, 4]]);
-        table.set_row_capacity(0);
-        table.set_column_capacity(0);
+        table.set_preferred_row_cnt(0);
+        table.set_preferred_col_cnt(0);
 
         assert_eq!(table.row_cnt(), 0);
         assert_eq!(table.col_cnt(), 0);
@@ -450,8 +458,8 @@ mod tests {
     #[should_panic]
     fn index_by_row_and_column_should_respect_virtual_boundaries() {
         let mut table = FixedColumnTable::from(vec![[1, 2, 3]]);
-        table.set_row_capacity(0);
-        table.set_column_capacity(0);
+        table.set_preferred_row_cnt(0);
+        table.set_preferred_col_cnt(0);
 
         // Will cause panic because of virtual boundary reached
         let _ = table[(0, 0)];
@@ -479,8 +487,8 @@ mod tests {
     #[should_panic]
     fn index_mut_by_row_and_column_should_respect_virtual_boundaries() {
         let mut table = FixedColumnTable::from(vec![[1, 2, 3]]);
-        table.set_row_capacity(0);
-        table.set_column_capacity(0);
+        table.set_preferred_row_cnt(0);
+        table.set_preferred_col_cnt(0);
 
         // Will cause panic because of virtual boundary reached
         table[(0, 0)] = 999;
@@ -534,7 +542,7 @@ mod tests {
         let mut table = FixedColumnTable::from([["a", "b", "c", "g"], ["d", "e", "f", "h"]]);
 
         // Shrink our capacity from the starting maximum so we can add a column
-        table.set_column_capacity(3);
+        table.set_preferred_col_cnt(3);
 
         table.insert_column(3, ["x", "y"].iter().copied());
 
@@ -573,7 +581,7 @@ mod tests {
         let mut table = FixedColumnTable::from([["a", "b", "c"], ["d", "e", "f"]]);
 
         // Shrink our capacity from the starting maximum so we can add a column
-        table.set_column_capacity(2);
+        table.set_preferred_col_cnt(2);
 
         table.push_column(["g", "h"].iter().copied());