+++ memtable-core/src/impls/fixed.rs
@@ -1,4 +1,4 @@
-use crate::{iter::*, list::*, utils, Position, Table};
+use crate::{iter::*, list::*, utils, Capacity, Position, Table};
 use core::{
     cmp,
     iter::FromIterator,
@@ -76,6 +76,14 @@ impl<T: Default, const ROW: usize, const COL: usize> Table for FixedTable<T, ROW
     type Row = FixedList<Self::Data, COL>;
     type Column = FixedList<Self::Data, ROW>;
 
+    fn max_row_capacity(&self) -> Capacity {
+        Capacity::Limited(ROW)
+    }
+
+    fn max_column_capacity(&self) -> Capacity {
+        Capacity::Limited(COL)
+    }
+
     fn row_cnt(&self) -> usize {
         self.row_cnt
     }
@@ -149,8 +157,8 @@ impl<T: Default, const ROW: usize, const COL: usize> Table for FixedTable<T, ROW
     ///
     /// If you want to remove the cells that are no longer within capacity,
     /// call [`Self::truncate`], which will reset them to their default value.
-    fn set_row_capacity(&mut self, capacity: usize) {
-        self.row_cnt = cmp::min(capacity, ROW);
+    fn set_preferred_row_cnt(&mut self, cnt: usize) {
+        self.row_cnt = cmp::min(cnt, ROW);
     }
 
     /// Will adjust the internal column count tracker to the specified capacity,
@@ -162,8 +170,8 @@ impl<T: Default, const ROW: usize, const COL: usize> Table for FixedTable<T, ROW
     ///
     /// If you want to remove the cells that are no longer within capacity,
     /// call [`Self::truncate`], which will reset them to their default value.
-    fn set_column_capacity(&mut self, capacity: usize) {
-        self.col_cnt = cmp::min(capacity, COL);
+    fn set_preferred_col_cnt(&mut self, cnt: usize) {
+        self.col_cnt = cmp::min(cnt, COL);
     }
 }
 
@@ -294,12 +302,12 @@ mod tests {
     fn row_cnt_should_be_adjustable_up_to_const_max() {
         let mut table: FixedTable<usize, 0, 0> = FixedTable::new();
         assert_eq!(table.row_cnt(), 0);
-        table.set_row_capacity(1);
+        table.set_preferred_row_cnt(1);
         assert_eq!(table.row_cnt(), 0);
 
         let mut table: FixedTable<usize, 4, 0> = FixedTable::new();
         assert_eq!(table.row_cnt(), 0);
-        table.set_row_capacity(5);
+        table.set_preferred_row_cnt(5);
         assert_eq!(table.row_cnt(), 4);
     }
 
@@ -307,12 +315,12 @@ mod tests {
     fn col_cnt_should_be_adjustable_up_to_const_max() {
         let mut table: FixedTable<usize, 0, 0> = FixedTable::new();
         assert_eq!(table.col_cnt(), 0);
-        table.set_column_capacity(1);
+        table.set_preferred_col_cnt(1);
         assert_eq!(table.col_cnt(), 0);
 
         let mut table: FixedTable<usize, 0, 4> = FixedTable::new();
         assert_eq!(table.col_cnt(), 0);
-        table.set_column_capacity(5);
+        table.set_preferred_col_cnt(5);
         assert_eq!(table.col_cnt(), 4);
     }
 
@@ -335,16 +343,16 @@ mod tests {
         assert_eq!(table.col_cnt(), 2);
 
         // If we change the capacity to be smaller, cell should respect that
-        table.set_row_capacity(1);
-        table.set_column_capacity(1);
+        table.set_preferred_row_cnt(1);
+        table.set_preferred_col_cnt(1);
         assert_eq!(table.cell(0, 0).as_deref(), Some(&"a"));
         assert_eq!(table.cell(0, 1).as_deref(), None);
         assert_eq!(table.cell(1, 0).as_deref(), None);
         assert_eq!(table.cell(1, 1).as_deref(), None);
 
         // Capacity changes don't actually overwrite anything
-        table.set_row_capacity(2);
-        table.set_column_capacity(2);
+        table.set_preferred_row_cnt(2);
+        table.set_preferred_col_cnt(2);
         assert_eq!(table.cell(0, 0).as_deref(), Some(&"a"));
         assert_eq!(table.cell(0, 1).as_deref(), Some(&"b"));
         assert_eq!(table.cell(1, 0).as_deref(), Some(&"c"));
@@ -373,8 +381,8 @@ mod tests {
         assert_eq!(table.col_cnt(), 2);
 
         // If we change the capacity to be smaller, mut_cell should respect that
-        table.set_row_capacity(1);
-        table.set_column_capacity(1);
+        table.set_preferred_row_cnt(1);
+        table.set_preferred_col_cnt(1);
         assert!(table.mut_cell(0, 0).is_some());
         assert!(table.mut_cell(0, 1).is_none());
         assert!(table.mut_cell(1, 0).is_none());
@@ -431,8 +439,8 @@ mod tests {
     #[test]
     fn remove_cell_should_respect_virtual_boundaries() {
         let mut table = FixedTable::from([[1, 2], [3, 4]]);
-        table.set_row_capacity(0);
-        table.set_column_capacity(0);
+        table.set_preferred_row_cnt(0);
+        table.set_preferred_col_cnt(0);
 
         assert_eq!(table.row_cnt(), 0);
         assert_eq!(table.col_cnt(), 0);
@@ -449,8 +457,8 @@ mod tests {
     #[should_panic]
     fn index_by_row_and_column_should_respect_virtual_boundaries() {
         let mut table = FixedTable::from([[1, 2, 3]]);
-        table.set_row_capacity(0);
-        table.set_column_capacity(0);
+        table.set_preferred_row_cnt(0);
+        table.set_preferred_col_cnt(0);
 
         // Will cause panic because of virtual boundary reached
         let _ = table[(0, 0)];
@@ -478,8 +486,8 @@ mod tests {
     #[should_panic]
     fn index_mut_by_row_and_column_should_respect_virtual_boundaries() {
         let mut table = FixedTable::from([[1, 2, 3]]);
-        table.set_row_capacity(0);
-        table.set_column_capacity(0);
+        table.set_preferred_row_cnt(0);
+        table.set_preferred_col_cnt(0);
 
         // Will cause panic because of virtual boundary reached
         table[(0, 0)] = 999;
@@ -497,7 +505,7 @@ mod tests {
         let mut table = FixedTable::from([["a", "b", "c"], ["d", "e", "f"], ["g", "h", "i"]]);
 
         // Shrink our capacity from the starting maximum so we can add a row
-        table.set_row_capacity(2);
+        table.set_preferred_row_cnt(2);
 
         table.insert_row(2, ["x", "y", "z"].iter().copied());
 
@@ -536,7 +544,7 @@ mod tests {
         let mut table = FixedTable::from([["a", "b", "c"], ["d", "e", "f"]]);
 
         // Shrink our capacity from the starting maximum so we can add a row
-        table.set_row_capacity(1);
+        table.set_preferred_row_cnt(1);
 
         table.push_row(["g", "h", "i"].iter().copied());
 
@@ -557,7 +565,7 @@ mod tests {
         let mut table = FixedTable::from([["a", "b", "c", "g"], ["d", "e", "f", "h"]]);
 
         // Shrink our capacity from the starting maximum so we can add a column
-        table.set_column_capacity(3);
+        table.set_preferred_col_cnt(3);
 
         table.insert_column(3, ["x", "y"].iter().copied());
 
@@ -596,7 +604,7 @@ mod tests {
         let mut table = FixedTable::from([["a", "b", "c"], ["d", "e", "f"]]);
 
         // Shrink our capacity from the starting maximum so we can add a column
-        table.set_column_capacity(2);
+        table.set_preferred_col_cnt(2);
 
         table.push_column(["g", "h"].iter().copied());